requirements.txt
# Core data handling
pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0

# API and web requests
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0

# Data analysis and forecasting
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.0
xgboost>=2.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Data validation
pydantic>=2.0.0
